cleanup before pullrequest
